Moved methods related to making wrestlers from templates into a concern to clear up the controller.
